+++ core/interface/orchestrator_action/src/lib.rs
@@ -1,10 +1,12 @@
 pub use replicante_models_core::actions::orchestrator::OrchestratorActionMetadata;
 
-mod error;
+use replicante_models_core::actions::orchestrator::OrchestratorActionState;
+
+pub mod errors;
 mod registry;
 mod traits;
 
-pub use self::error::ActionAlreadyRegistered;
+pub use self::errors::ActionAlreadyRegistered;
 pub use self::registry::OrchestratorActionRegistry;
 pub use self::registry::OrchestratorActionRegistryBuilder;
 pub use self::registry::OrchestratorActionRegistryEntry;
@@ -12,3 +14,9 @@ pub use self::traits::OrchestratorAction;
 
 #[cfg(feature = "test-api")]
 pub use self::registry::TestRegistryClearGuard;
+
+/// Store changes made to an `OrchestratorAction` model as a result of progressing the action.
+pub struct ProgressChanges {
+    pub state: OrchestratorActionState,
+    pub state_payload: Option<serde_json::Value>,
+}